fix: refuse a template send naming both sources
Addresses the CodeRabbit review. A send carrying both TemplateName and TemplateContent rendered the inline wording while recording the stored template's name, filing the message under a template it never came from. Which of the two real SES prefers is not something this simulator knows, so the combination is now refused rather than guessed at. The docs also now say that TemplateData is only checked when a send carries one, since an omitted one is accepted.
